requirements.txt
six
pyserial
numpy